refactor(coderd/x/chatd): extract OpenAI Responses chain mode helpers to chatopenai
