+++ coderd/x/chatd/chatd_internal_test.go
@@ -9,7 +9,6 @@ import (
 	"time"
 
 	"charm.land/fantasy"
-	fantasyopenai "charm.land/fantasy/providers/openai"
 	"github.com/google/uuid"
 	"github.com/sqlc-dev/pqtype"
 	"github.com/stretchr/testify/require"
@@ -24,7 +23,6 @@ import (
 	coderdpubsub "github.com/coder/coder/v2/coderd/pubsub"
 	"github.com/coder/coder/v2/coderd/x/chatd/chaterror"
 	"github.com/coder/coder/v2/coderd/x/chatd/chatloop"
-	"github.com/coder/coder/v2/coderd/x/chatd/chatprompt"
 	"github.com/coder/coder/v2/coderd/x/chatd/chatprovider"
 	"github.com/coder/coder/v2/coderd/x/chatd/chattest"
 	"github.com/coder/coder/v2/coderd/x/chatd/chattool"
@@ -2875,618 +2873,6 @@ func TestSelectSkillMetasForInstructionRefresh(t *testing.T) {
 		)
 		require.Equal(t, persisted, got)
 	})
-}
-
-func TestResolveChainModeIgnoresSkillOnlySentinelMessages(t *testing.T) {
-	t.Parallel()
-
-	modelConfigID := uuid.New()
-	assistant := database.ChatMessage{
-		Role:               database.ChatMessageRoleAssistant,
-		ProviderResponseID: sql.NullString{String: "resp-123", Valid: true},
-		ModelConfigID:      uuid.NullUUID{UUID: modelConfigID, Valid: true},
-	}
-	skillOnly := chatMessageWithParts([]codersdk.ChatMessagePart{
-		{
-			Type:            codersdk.ChatMessagePartTypeContextFile,
-			ContextFilePath: AgentChatContextSentinelPath,
-			ContextFileAgentID: uuid.NullUUID{
-				UUID:  uuid.New(),
-				Valid: true,
-			},
-		},
-		{
-			Type:      codersdk.ChatMessagePartTypeSkill,
-			SkillName: "repo-helper",
-			SkillDir:  "/skills/repo-helper",
-		},
-	})
-	skillOnly.Role = database.ChatMessageRoleUser
-	user := chatMessageWithParts([]codersdk.ChatMessagePart{{
-		Type: codersdk.ChatMessagePartTypeText,
-		Text: "latest user message",
-	}})
-	user.Role = database.ChatMessageRoleUser
-
-	got := resolveChainMode([]database.ChatMessage{assistant, skillOnly, user})
-	require.Equal(t, "resp-123", got.previousResponseID)
-	require.Equal(t, modelConfigID, got.modelConfigID)
-	require.Equal(t, 2, got.trailingUserCount)
-	require.Equal(t, 1, got.contributingTrailingUserCount)
-}
-
-func TestResolveChainMode_BlocksOnUnresolvedLocalToolCall(t *testing.T) {
-	t.Parallel()
-
-	modelConfigID := uuid.New()
-	toolCall := codersdk.ChatMessageToolCall(
-		"call-local",
-		"read_file",
-		json.RawMessage(`{"path":"main.go"}`),
-	)
-
-	chainInfo := resolveChainMode([]database.ChatMessage{
-		chainModeSystemMessage(),
-		chainModeUserMessage("prior user message"),
-		chainModeAssistantMessage(modelConfigID, []codersdk.ChatMessagePart{toolCall}),
-		chainModeUserMessage("latest user message"),
-	})
-
-	require.Equal(t, "resp-123", chainInfo.previousResponseID)
-	require.True(t, chainInfo.hasUnresolvedLocalToolCalls)
-	require.False(t, shouldActivateChainMode(
-		chainModeProviderOptions(),
-		chainInfo,
-		modelConfigID,
-		false,
-	))
-}
-
-func TestResolveChainMode_BlocksWhenAssistantContentCannotParse(t *testing.T) {
-	t.Parallel()
-
-	modelConfigID := uuid.New()
-	chainInfo := resolveChainMode([]database.ChatMessage{
-		chainModeSystemMessage(),
-		chainModeUserMessage("prior user message"),
-		chainModeCorruptAssistantMessage(modelConfigID),
-		chainModeUserMessage("latest user message"),
-	})
-
-	require.Equal(t, "resp-123", chainInfo.previousResponseID)
-	require.True(t, chainInfo.hasUnresolvedLocalToolCalls)
-	require.False(t, shouldActivateChainMode(
-		chainModeProviderOptions(),
-		chainInfo,
-		modelConfigID,
-		false,
-	))
-}
-
-func TestResolveChainMode_BlocksWhenToolContentCannotParse(t *testing.T) {
-	t.Parallel()
-
-	modelConfigID := uuid.New()
-	toolCall := codersdk.ChatMessageToolCall(
-		"call-local",
-		"read_file",
-		json.RawMessage(`{"path":"main.go"}`),
-	)
-
-	chainInfo := resolveChainMode([]database.ChatMessage{
-		chainModeSystemMessage(),
-		chainModeUserMessage("prior user message"),
-		chainModeAssistantMessage(modelConfigID, []codersdk.ChatMessagePart{toolCall}),
-		chainModeCorruptToolMessage(),
-		chainModeUserMessage("latest user message"),
-	})
-
-	require.Equal(t, "resp-123", chainInfo.previousResponseID)
-	require.True(t, chainInfo.hasUnresolvedLocalToolCalls)
-	require.False(t, shouldActivateChainMode(
-		chainModeProviderOptions(),
-		chainInfo,
-		modelConfigID,
-		false,
-	))
-}
-
-func TestResolveChainMode_AllowsProviderExecutedOnly(t *testing.T) {
-	t.Parallel()
-
-	modelConfigID := uuid.New()
-	toolCall := codersdk.ChatMessageToolCall(
-		"call-web-search",
-		"web_search",
-		json.RawMessage(`{"query":"coder docs"}`),
-	)
-	toolCall.ProviderExecuted = true
-
-	chainInfo := resolveChainMode([]database.ChatMessage{
-		chainModeSystemMessage(),
-		chainModeUserMessage("prior user message"),
-		chainModeAssistantMessage(modelConfigID, []codersdk.ChatMessagePart{toolCall}),
-		chainModeUserMessage("latest user message"),
-	})
-
-	require.Equal(t, "resp-123", chainInfo.previousResponseID)
-	require.False(t, chainInfo.hasUnresolvedLocalToolCalls)
-	require.False(t, chainInfo.providerMissingToolResults)
-	require.True(t, shouldActivateChainMode(
-		chainModeProviderOptions(),
-		chainInfo,
-		modelConfigID,
-		false,
-	))
-}
-
-func TestResolveChainMode_BlocksOnMixedProviderExecutedAndUnresolvedLocalCall(t *testing.T) {
-	t.Parallel()
-
-	modelConfigID := uuid.New()
-	providerCall := codersdk.ChatMessageToolCall(
-		"call-web-search",
-		"web_search",
-		json.RawMessage(`{"query":"coder docs"}`),
-	)
-	providerCall.ProviderExecuted = true
-	localCall := codersdk.ChatMessageToolCall(
-		"call-local",
-		"read_file",
-		json.RawMessage(`{"path":"main.go"}`),
-	)
-
-	chainInfo := resolveChainMode([]database.ChatMessage{
-		chainModeSystemMessage(),
-		chainModeUserMessage("prior user message"),
-		chainModeAssistantMessage(
-			modelConfigID,
-			[]codersdk.ChatMessagePart{providerCall, localCall},
-		),
-		chainModeUserMessage("latest user message"),
-	})
-
-	require.Equal(t, "resp-123", chainInfo.previousResponseID)
-	require.True(t, chainInfo.hasUnresolvedLocalToolCalls)
-	require.False(t, shouldActivateChainMode(
-		chainModeProviderOptions(),
-		chainInfo,
-		modelConfigID,
-		false,
-	))
-}
-
-func TestResolveChainMode_AllowsResolvedLocalCall(t *testing.T) {
-	t.Parallel()
-
-	modelConfigID := uuid.New()
-	toolCall := codersdk.ChatMessageToolCall(
-		"call-local",
-		"read_file",
-		json.RawMessage(`{"path":"main.go"}`),
-	)
-	toolResult := codersdk.ChatMessageToolResult(
-		"call-local",
-		"read_file",
-		json.RawMessage(`{"ok":true}`),
-		false,
-		false,
-	)
-
-	// A follow-up assistant after the tool result confirms the
-	// result was sent back to the provider. Chain mode should
-	// activate from the follow-up assistant's response ID.
-	// Use a distinct response ID on the follow-up assistant
-	// so the assertion verifies resolveChainMode selects the
-	// follow-up (last assistant), not the original tool-caller.
-	followUp := chainModeAssistantMessage(modelConfigID, nil)
-	followUp.ProviderResponseID = sql.NullString{String: "resp-follow-up", Valid: true}
-
-	chainInfo := resolveChainMode([]database.ChatMessage{
-		chainModeSystemMessage(),
-		chainModeUserMessage("prior user message"),
-		chainModeAssistantMessage(modelConfigID, []codersdk.ChatMessagePart{toolCall}),
-		chainModeToolMessage([]codersdk.ChatMessagePart{toolResult}),
-		followUp,
-		chainModeUserMessage("latest user message"),
-	})
-
-	require.Equal(t, "resp-follow-up", chainInfo.previousResponseID)
-	require.False(t, chainInfo.hasUnresolvedLocalToolCalls)
-	require.False(t, chainInfo.providerMissingToolResults)
-	require.True(t, shouldActivateChainMode(
-		chainModeProviderOptions(),
-		chainInfo,
-		modelConfigID,
-		false,
-	))
-}
-
-func TestResolveChainMode_BlocksOnMixedResolvedAndUnresolved(t *testing.T) {
-	t.Parallel()
-
-	modelConfigID := uuid.New()
-	firstCall := codersdk.ChatMessageToolCall(
-		"call-first",
-		"read_file",
-		json.RawMessage(`{"path":"main.go"}`),
-	)
-	secondCall := codersdk.ChatMessageToolCall(
-		"call-second",
-		"read_file",
-		json.RawMessage(`{"path":"README.md"}`),
-	)
-	toolResult := codersdk.ChatMessageToolResult(
-		"call-first",
-		"read_file",
-		json.RawMessage(`{"ok":true}`),
-		false,
-		false,
-	)
-
-	chainInfo := resolveChainMode([]database.ChatMessage{
-		chainModeSystemMessage(),
-		chainModeUserMessage("prior user message"),
-		chainModeAssistantMessage(
-			modelConfigID,
-			[]codersdk.ChatMessagePart{firstCall, secondCall},
-		),
-		chainModeToolMessage([]codersdk.ChatMessagePart{toolResult}),
-		chainModeUserMessage("latest user message"),
-	})
-
-	require.Equal(t, "resp-123", chainInfo.previousResponseID)
-	require.True(t, chainInfo.hasUnresolvedLocalToolCalls)
-	require.False(t, shouldActivateChainMode(
-		chainModeProviderOptions(),
-		chainInfo,
-		modelConfigID,
-		false,
-	))
-}
-
-// Tests for providerMissingToolResults detection.
-// These cover the StopAfterTool + chain mode desync bug where local
-// tool results exist in the DB but were never sent back to the
-// provider, leaving an unresolved function_call in the stored chain.
-
-func TestResolveChainMode_BlocksWhenToolResultNeverSentToProvider(t *testing.T) {
-	t.Parallel()
-
-	modelConfigID := uuid.New()
-	toolCall := codersdk.ChatMessageToolCall(
-		"call-local",
-		"propose_plan",
-		json.RawMessage(`{"path":"plan.md"}`),
-	)
-	toolResult := codersdk.ChatMessageToolResult(
-		"call-local",
-		"propose_plan",
-		json.RawMessage(`{"ok":true}`),
-		false,
-		false,
-	)
-
-	chainInfo := resolveChainMode([]database.ChatMessage{
-		chainModeSystemMessage(),
-		chainModeUserMessage("make a plan"),
-		chainModeAssistantMessage(modelConfigID, []codersdk.ChatMessagePart{toolCall}),
-		chainModeToolMessage([]codersdk.ChatMessagePart{toolResult}),
-		// No follow-up assistant: StopAfterTool fired, tool result
-		// was persisted locally but never sent back to the provider.
-		chainModeUserMessage("implement the plan"),
-	})
-
-	require.Equal(t, "resp-123", chainInfo.previousResponseID)
-	// Local tool calls are resolved (result exists in DB).
-	require.False(t, chainInfo.hasUnresolvedLocalToolCalls)
-	// But the provider never received the result.
-	require.True(t, chainInfo.providerMissingToolResults)
-	// Chain mode must NOT activate.
-	require.False(t, shouldActivateChainMode(
-		chainModeProviderOptions(),
-		chainInfo,
-		modelConfigID,
-		false,
-	))
-}
-
-func TestResolveChainMode_BlocksProviderMissingWithMultipleToolCalls(t *testing.T) {
-	t.Parallel()
-
-	modelConfigID := uuid.New()
-	call1 := codersdk.ChatMessageToolCall(
-		"call-1", "propose_plan",
-		json.RawMessage(`{"path":"plan.md"}`),
-	)
-	call2 := codersdk.ChatMessageToolCall(
-		"call-2", "write_file",
-		json.RawMessage(`{"path":"foo.go"}`),
-	)
-	result1 := codersdk.ChatMessageToolResult(
-		"call-1", "propose_plan",
-		json.RawMessage(`{"ok":true}`), false, false,
-	)
-	result2 := codersdk.ChatMessageToolResult(
-		"call-2", "write_file",
-		json.RawMessage(`{"ok":true}`), false, false,
-	)
-
-	chainInfo := resolveChainMode([]database.ChatMessage{
-		chainModeSystemMessage(),
-		chainModeUserMessage("do it"),
-		chainModeAssistantMessage(modelConfigID, []codersdk.ChatMessagePart{call1, call2}),
-		chainModeToolMessage([]codersdk.ChatMessagePart{result1, result2}),
-		chainModeUserMessage("next"),
-	})
-
-	require.False(t, chainInfo.hasUnresolvedLocalToolCalls)
-	require.True(t, chainInfo.providerMissingToolResults)
-	require.False(t, shouldActivateChainMode(
-		chainModeProviderOptions(), chainInfo, modelConfigID, false,
-	))
-}
-
-func TestResolveChainMode_AllowsWhenNoToolCalls(t *testing.T) {
-	t.Parallel()
-
-	modelConfigID := uuid.New()
-
-	chainInfo := resolveChainMode([]database.ChatMessage{
-		chainModeSystemMessage(),
-		chainModeUserMessage("hello"),
-		chainModeAssistantMessage(modelConfigID, nil),
-		chainModeUserMessage("thanks"),
-	})
-
-	require.Equal(t, "resp-123", chainInfo.previousResponseID)
-	require.False(t, chainInfo.hasUnresolvedLocalToolCalls)
-	require.False(t, chainInfo.providerMissingToolResults)
-	require.True(t, shouldActivateChainMode(
-		chainModeProviderOptions(), chainInfo, modelConfigID, false,
-	))
-}
-
-func chainModeProviderOptions() fantasy.ProviderOptions {
-	store := true
-	return fantasy.ProviderOptions{
-		fantasyopenai.Name: &fantasyopenai.ResponsesProviderOptions{
-			Store: &store,
-		},
-	}
-}
-
-func chainModeSystemMessage() database.ChatMessage {
-	return database.ChatMessage{Role: database.ChatMessageRoleSystem}
-}
-
-func chainModeUserMessage(text string) database.ChatMessage {
-	msg := chatMessageWithParts([]codersdk.ChatMessagePart{
-		codersdk.ChatMessageText(text),
-	})
-	msg.Role = database.ChatMessageRoleUser
-	return msg
-}
-
-func chainModeAssistantMessage(
-	modelConfigID uuid.UUID,
-	parts []codersdk.ChatMessagePart,
-) database.ChatMessage {
-	msg := chatMessageWithParts(parts)
-	msg.Role = database.ChatMessageRoleAssistant
-	msg.ProviderResponseID = sql.NullString{String: "resp-123", Valid: true}
-	msg.ModelConfigID = uuid.NullUUID{UUID: modelConfigID, Valid: true}
-	return msg
-}
-
-func chainModeCorruptAssistantMessage(modelConfigID uuid.UUID) database.ChatMessage {
-	return database.ChatMessage{
-		Role:               database.ChatMessageRoleAssistant,
-		ProviderResponseID: sql.NullString{String: "resp-123", Valid: true},
-		ModelConfigID:      uuid.NullUUID{UUID: modelConfigID, Valid: true},
-		Content: pqtype.NullRawMessage{
-			RawMessage: []byte("not json"),
-			Valid:      true,
-		},
-		ContentVersion: chatprompt.CurrentContentVersion,
-	}
-}
-
-func chainModeCorruptToolMessage() database.ChatMessage {
-	return database.ChatMessage{
-		Role: database.ChatMessageRoleTool,
-		Content: pqtype.NullRawMessage{
-			RawMessage: []byte("not json"),
-			Valid:      true,
-		},
-		ContentVersion: chatprompt.CurrentContentVersion,
-	}
-}
-
-func chainModeToolMessage(parts []codersdk.ChatMessagePart) database.ChatMessage {
-	msg := chatMessageWithParts(parts)
-	msg.Role = database.ChatMessageRoleTool
-	return msg
-}
-
-func TestFilterPromptForChainModeKeepsContributingUsersAcrossSkippedSentinelTurns(t *testing.T) {
-	t.Parallel()
-
-	modelConfigID := uuid.New()
-	priorUser := chatMessageWithParts([]codersdk.ChatMessagePart{{
-		Type: codersdk.ChatMessagePartTypeText,
-		Text: "prior user message",
-	}})
-	priorUser.Role = database.ChatMessageRoleUser
-	assistant := database.ChatMessage{
-		Role:               database.ChatMessageRoleAssistant,
-		ProviderResponseID: sql.NullString{String: "resp-123", Valid: true},
-		ModelConfigID:      uuid.NullUUID{UUID: modelConfigID, Valid: true},
-	}
-	firstTrailingUser := chatMessageWithParts([]codersdk.ChatMessagePart{{
-		Type: codersdk.ChatMessagePartTypeText,
-		Text: "first trailing user",
-	}})
-	firstTrailingUser.Role = database.ChatMessageRoleUser
-	skillOnly := chatMessageWithParts([]codersdk.ChatMessagePart{
-		{
-			Type:            codersdk.ChatMessagePartTypeContextFile,
-			ContextFilePath: AgentChatContextSentinelPath,
-			ContextFileAgentID: uuid.NullUUID{
-				UUID:  uuid.New(),
-				Valid: true,
-			},
-		},
-		{
-			Type:      codersdk.ChatMessagePartTypeSkill,
-			SkillName: "repo-helper",
-			SkillDir:  "/skills/repo-helper",
-		},
-	})
-	skillOnly.Role = database.ChatMessageRoleUser
-	lastTrailingUser := chatMessageWithParts([]codersdk.ChatMessagePart{{
-		Type: codersdk.ChatMessagePartTypeText,
-		Text: "last trailing user",
-	}})
-	lastTrailingUser.Role = database.ChatMessageRoleUser
-
-	chainInfo := resolveChainMode([]database.ChatMessage{
-		priorUser,
-		assistant,
-		firstTrailingUser,
-		skillOnly,
-		lastTrailingUser,
-	})
-	require.Equal(t, 3, chainInfo.trailingUserCount)
-	require.Equal(t, 2, chainInfo.contributingTrailingUserCount)
-
-	prompt := []fantasy.Message{
-		{
-			Role: fantasy.MessageRoleSystem,
-			Content: []fantasy.MessagePart{
-				fantasy.TextPart{Text: "system instruction"},
-			},
-		},
-		{
-			Role: fantasy.MessageRoleUser,
-			Content: []fantasy.MessagePart{
-				fantasy.TextPart{Text: "prior user message"},
-			},
-		},
-		{
-			Role: fantasy.MessageRoleAssistant,
-			Content: []fantasy.MessagePart{
-				fantasy.TextPart{Text: "assistant reply"},
-			},
-		},
-		{
-			Role: fantasy.MessageRoleUser,
-			Content: []fantasy.MessagePart{
-				fantasy.TextPart{Text: "first trailing user"},
-			},
-		},
-		{
-			Role: fantasy.MessageRoleUser,
-			Content: []fantasy.MessagePart{
-				fantasy.TextPart{Text: "last trailing user"},
-			},
-		},
-	}
-
-	got := filterPromptForChainMode(prompt, chainInfo)
-	require.Len(t, got, 3)
-	require.Equal(t, fantasy.MessageRoleSystem, got[0].Role)
-	require.Equal(t, fantasy.MessageRoleUser, got[1].Role)
-	require.Equal(t, fantasy.MessageRoleUser, got[2].Role)
-
-	firstPart, ok := fantasy.AsMessagePart[fantasy.TextPart](got[1].Content[0])
-	require.True(t, ok)
-	require.Equal(t, "first trailing user", firstPart.Text)
-	lastPart, ok := fantasy.AsMessagePart[fantasy.TextPart](got[2].Content[0])
-	require.True(t, ok)
-	require.Equal(t, "last trailing user", lastPart.Text)
-}
-
-func TestFilterPromptForChainModeUsesContributingTrailingUsers(t *testing.T) {
-	t.Parallel()
-
-	modelConfigID := uuid.New()
-	priorUser := chatMessageWithParts([]codersdk.ChatMessagePart{{
-		Type: codersdk.ChatMessagePartTypeText,
-		Text: "prior user message",
-	}})
-	priorUser.Role = database.ChatMessageRoleUser
-	assistant := database.ChatMessage{
-		Role:               database.ChatMessageRoleAssistant,
-		ProviderResponseID: sql.NullString{String: "resp-123", Valid: true},
-		ModelConfigID:      uuid.NullUUID{UUID: modelConfigID, Valid: true},
-	}
-	skillOnly := chatMessageWithParts([]codersdk.ChatMessagePart{
-		{
-			Type:            codersdk.ChatMessagePartTypeContextFile,
-			ContextFilePath: AgentChatContextSentinelPath,
-			ContextFileAgentID: uuid.NullUUID{
-				UUID:  uuid.New(),
-				Valid: true,
-			},
-		},
-		{
-			Type:      codersdk.ChatMessagePartTypeSkill,
-			SkillName: "repo-helper",
-			SkillDir:  "/skills/repo-helper",
-		},
-	})
-	skillOnly.Role = database.ChatMessageRoleUser
-	latestUser := chatMessageWithParts([]codersdk.ChatMessagePart{{
-		Type: codersdk.ChatMessagePartTypeText,
-		Text: "latest user message",
-	}})
-	latestUser.Role = database.ChatMessageRoleUser
-
-	chainInfo := resolveChainMode([]database.ChatMessage{
-		priorUser,
-		assistant,
-		skillOnly,
-		latestUser,
-	})
-	require.Equal(t, 2, chainInfo.trailingUserCount)
-	require.Equal(t, 1, chainInfo.contributingTrailingUserCount)
-
-	prompt := []fantasy.Message{
-		{
-			Role: fantasy.MessageRoleSystem,
-			Content: []fantasy.MessagePart{
-				fantasy.TextPart{Text: "system instruction"},
-			},
-		},
-		{
-			Role: fantasy.MessageRoleUser,
-			Content: []fantasy.MessagePart{
-				fantasy.TextPart{Text: "prior user message"},
-			},
-		},
-		{
-			Role: fantasy.MessageRoleAssistant,
-			Content: []fantasy.MessagePart{
-				fantasy.TextPart{Text: "assistant reply"},
-			},
-		},
-		{
-			Role: fantasy.MessageRoleUser,
-			Content: []fantasy.MessagePart{
-				fantasy.TextPart{Text: "latest user message"},
-			},
-		},
-	}
-
-	got := filterPromptForChainMode(prompt, chainInfo)
-	require.Len(t, got, 2)
-	require.Equal(t, fantasy.MessageRoleSystem, got[0].Role)
-	require.Equal(t, fantasy.MessageRoleUser, got[1].Role)
-
-	part, ok := fantasy.AsMessagePart[fantasy.TextPart](got[1].Content[0])
-	require.True(t, ok)
-	require.Equal(t, "latest user message", part.Text)
 }
 
 func chatMessageWithParts(parts []codersdk.ChatMessagePart) database.ChatMessage {
