+++ coderd/x/chatd/chatloop/chatloop.go
@@ -16,7 +16,6 @@ import (
 
 	"charm.land/fantasy"
 	fantasyanthropic "charm.land/fantasy/providers/anthropic"
-	fantasyopenai "charm.land/fantasy/providers/openai"
 	"charm.land/fantasy/schema"
 	"golang.org/x/xerrors"
 
@@ -24,6 +23,7 @@ import (
 	"github.com/coder/coder/v2/coderd/database/dbtime"
 	"github.com/coder/coder/v2/coderd/x/chatd/chatdebug"
 	"github.com/coder/coder/v2/coderd/x/chatd/chaterror"
+	"github.com/coder/coder/v2/coderd/x/chatd/chatopenai"
 	"github.com/coder/coder/v2/coderd/x/chatd/chatprompt"
 	"github.com/coder/coder/v2/coderd/x/chatd/chatretry"
 	"github.com/coder/coder/v2/coderd/x/chatd/chatsanitize"
@@ -512,7 +512,7 @@ func Run(ctx context.Context, opts RunOptions) error {
 				Content:             result.content,
 				Usage:               result.usage,
 				ContextLimit:        contextLimit,
-				ProviderResponseID:  extractOpenAIResponseIDIfStored(opts.ProviderOptions, result.providerMetadata),
+				ProviderResponseID:  chatopenai.ExtractResponseIDIfStored(opts.ProviderOptions, result.providerMetadata),
 				Runtime:             time.Since(stepStart),
 				ToolCallCreatedAt:   result.toolCallCreatedAt,
 				ToolResultCreatedAt: result.toolResultCreatedAt,
@@ -538,8 +538,8 @@ func Run(ctx context.Context, opts RunOptions) error {
 			// when previous_response_id is set, so we must leave chain
 			// mode and reload the full history before the next call.
 			stepMessages := result.toResponseMessages()
-			if hasPreviousResponseID(opts.ProviderOptions) {
-				clearPreviousResponseID(opts.ProviderOptions)
+			if chatopenai.HasPreviousResponseID(opts.ProviderOptions) {
+				chatopenai.ClearPreviousResponseID(opts.ProviderOptions)
 				if opts.DisableChainMode != nil {
 					opts.DisableChainMode()
 				}
@@ -1233,7 +1233,7 @@ func persistPendingDynamicStep(
 		Content:                 result.content,
 		Usage:                   result.usage,
 		ContextLimit:            contextLimit,
-		ProviderResponseID:      extractOpenAIResponseIDIfStored(opts.ProviderOptions, result.providerMetadata),
+		ProviderResponseID:      chatopenai.ExtractResponseIDIfStored(opts.ProviderOptions, result.providerMetadata),
 		Runtime:                 time.Since(stepStart),
 		PendingDynamicToolCalls: pending,
 	}); err != nil {
@@ -1707,85 +1707,6 @@ func addAnthropicPromptCaching(messages []fantasy.Message) {
 			messages[i].ProviderOptions = providerOption
 		}
 	}
-}
-
-// hasPreviousResponseID checks whether the provider options contain
-// an OpenAI Responses entry with a non-empty PreviousResponseID.
-func hasPreviousResponseID(providerOptions fantasy.ProviderOptions) bool {
-	if providerOptions == nil {
-		return false
-	}
-
-	for _, entry := range providerOptions {
-		if options, ok := entry.(*fantasyopenai.ResponsesProviderOptions); ok {
-			return options.PreviousResponseID != nil &&
-				*options.PreviousResponseID != ""
-		}
-	}
-
-	return false
-}
-
-// clearPreviousResponseID removes PreviousResponseID from the OpenAI
-// Responses provider options entry, if present.
-func clearPreviousResponseID(providerOptions fantasy.ProviderOptions) {
-	if providerOptions == nil {
-		return
-	}
-
-	for _, entry := range providerOptions {
-		if options, ok := entry.(*fantasyopenai.ResponsesProviderOptions); ok {
-			options.PreviousResponseID = nil
-		}
-	}
-}
-
-// extractOpenAIResponseID extracts the OpenAI Responses API response
-// ID from provider metadata. Returns an empty string if no OpenAI
-// Responses metadata is present.
-func extractOpenAIResponseID(metadata fantasy.ProviderMetadata) string {
-	if len(metadata) == 0 {
-		return ""
-	}
-
-	for _, entry := range metadata {
-		if providerMetadata, ok := entry.(*fantasyopenai.ResponsesProviderMetadata); ok && providerMetadata != nil {
-			return providerMetadata.ResponseID
-		}
-	}
-
-	return ""
-}
-
-// extractOpenAIResponseIDIfStored returns the OpenAI response ID
-// only when the provider options indicate store=true. Response IDs
-// from store=false turns are not persisted server-side and cannot
-// be used for chaining.
-func extractOpenAIResponseIDIfStored(
-	providerOptions fantasy.ProviderOptions,
-	metadata fantasy.ProviderMetadata,
-) string {
-	if !isResponsesStoreEnabled(providerOptions) {
-		return ""
-	}
-
-	return extractOpenAIResponseID(metadata)
-}
-
-// isResponsesStoreEnabled checks whether the OpenAI Responses
-// provider options explicitly enable store=true.
-func isResponsesStoreEnabled(providerOptions fantasy.ProviderOptions) bool {
-	if providerOptions == nil {
-		return false
-	}
-
-	for _, entry := range providerOptions {
-		if options, ok := entry.(*fantasyopenai.ResponsesProviderOptions); ok {
-			return options.Store != nil && *options.Store
-		}
-	}
-
-	return false
 }
 
 // recordToolResultTimestamp lazily initializes the
